chore: separate ssz_proof to a new crate
