+++ beacon_node/http_api/Cargo.toml
@@ -6,7 +6,6 @@ edition = { workspace = true }
 autotests = false                               # using a single test binary compiles faster
 
 [dependencies]
-ssz_rs = { git = "https://github.com/ralexstokes/ssz-rs", rev = "84ef2b71aa004f6767420badb42c902ad56b8b72" }
 warp = { workspace = true }
 serde = { workspace = true }
 tokio = { workspace = true }
@@ -18,6 +17,7 @@ eth2 = { workspace = true }
 slog = { workspace = true }
 network = { workspace = true }
 lighthouse_network = { workspace = true }
+ssz_proof = { workspace = true }
 eth1 = { workspace = true }
 state_processing = { workspace = true }
 lighthouse_version = { workspace = true }
