+++ consensus/ssz_proof/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "ssz_proof"
+version = "0.2.0"
+edition = { workspace = true }
+
+[dependencies]
+ssz_rs = { git = "https://github.com/ralexstokes/ssz-rs", rev = "84ef2b71aa004f6767420badb42c902ad56b8b72" }
+ssz_rs_derive = { git = "https://github.com/ralexstokes/ssz-rs", rev = "84ef2b71aa004f6767420badb42c902ad56b8b72" }
+ethereum_consensus = { git = "https://github.com/ralexstokes/ethereum-consensus", package = "ethereum-consensus", branch = "main" }
+serde = { workspace = true }
+types = { workspace = true }
+hex = { workspace = true }
+ssz_types = { workspace = true, features = ["arbitrary"] }
+tree_hash = { workspace = true, features = ["arbitrary"] }
+tree_hash_derive = { workspace = true }
+derivative = { workspace = true }
+arbitrary = { workspace = true, features = ["derive"] }
+ethereum_serde_utils = { workspace = true }
+superstruct = { workspace = true }
+metastruct = "0.1.0"
+serde_json = { workspace = true }
+milhouse = { workspace = true }
