+++ consensus/types/src/lib.rs
@@ -25,7 +25,6 @@ pub mod beacon_block_body;
 pub mod beacon_block_header;
 pub mod beacon_committee;
 pub mod beacon_state;
-pub mod beacon_state_summary;
 pub mod bls_to_execution_change;
 pub mod builder_bid;
 pub mod chain_spec;
